Move IFrameNode to separate source file and add tests.
